+++ ACL/src/ACL.Business/Application/Interfaces/Repositories/ICompanyModuleRepository.cs
@@ -10,7 +10,7 @@ namespace ACL.Business.Application.Interfaces.Repositories
         /// <inheritdoc/>
         List<CompanyModule>? All();
         /// <inheritdoc/>
-        CompanyModule? Find(ulong id);
+        CompanyModule? Find(uint id);
         /// <inheritdoc/>
         CompanyModule? Add(CompanyModule aclCompanyModule);
         /// <inheritdoc/>
@@ -18,7 +18,7 @@ namespace ACL.Business.Application.Interfaces.Repositories
         /// <inheritdoc/>
         CompanyModule? Delete(CompanyModule aclCompanyModule);
         /// <inheritdoc/>
-        CompanyModule? Delete(ulong id);
+        CompanyModule? Delete(uint id);
 
     }
 }
