+++ ACL/src/ACL.Business/Domain/Services/IUserGroupRoleService.cs
@@ -7,7 +7,7 @@ namespace ACL.Business.Domain.Services
     public interface IUserGroupRoleService : IUserGroupRoleRepository
     {
         /// <inheritdoc/>
-        ScopeResponse GetRolesByUserGroupId(ulong groupId);
+        ScopeResponse GetRolesByUserGroupId(uint groupId);
         /// <inheritdoc/>
         Task<ScopeResponse> Update(AclUserGroupRoleRequest userGroupRole);
     }
